refactor(content): extract image MIME/extension table to a leaf module
Counsel (codex) on B.1 flagged a boundary leak: the pure image-optimize
boundary imported extensionForType from admin-form (form/key policy), and the
corrupt-upload fallback echoed the raw source MIME casing. Move the accepted-MIME
gate + MIME<->extension table to a leaf image-types module both admin-form and
image-optimize import; normalize the source MIME once so the fallback content-type
is clean lowercase.
